Added types for `Money` to be understood as pure/immutable downstream
Note: this does *NOT* make `Money` per-se pass
any sort of internal static analysis, but
only declares a thin contract between
consumers of `Money` and its public
API, allowing downstream to rely on the
immutable invariants of the implementation
provided thus far.

Adding further static analysis and types
is (for now) out of scope, and the
patch limits itself to making the
`Money` type usable in other contexts
that are equally annotated as pure/immutable.