Stricter type checks around invalid operands, hunting down unused/redundant code
Since phpspec is no longer first-class citizen of the library, hereby we converted some
of the specs into unit tests, and ensured that all of the API of the library is not
unused (means there's at least another symbol or test relying on it).

In addition to that, operations such as mixed `float/int`, `string/int`, etc math
are now generally removed from allowed code.